***BE CAREFUL WITH THIS VERSION*** ***BE CAREFUL WITH THIS VERSION***
I'm trying to run the regression function and this version is attampts to understand why we don't have data for the regression, mostly invastigating formatted_dt.

It's an unfinished one and has one mistake spotted for sure- religious is NA in the formatting per wave.

The reason I'm committing is cuz I created a function that can take any data set and show how many NAs each of its field has- might be useful